+++ tridesclous/tests/test_metrics.py
@@ -8,13 +8,16 @@ from tridesclous.tests.testingtools import ON_CI_CLOUD
 import pytest
 
 import shutil
+import os
 
 
 def setup_module():
     setup_catalogue('test_metrics', dataset_name='olfactory_bulb')
 
 def teardown_module():
-    shutil.rmtree('test_metrics')
+    if not(os.environ.get('APPVEYOR') in ('true', 'True')):
+        # this fix appveyor teardown_module bug
+        shutil.rmtree('test_metrics')
     
     
 
@@ -26,6 +29,8 @@ def test_all_metrics():
     cc.compute_cluster_similarity()
     cc.compute_cluster_ratio_similarity()
     cc.compute_spike_silhouette()
+
+    
 
 
 @pytest.mark.skipif(ON_CI_CLOUD, reason='ON_CI_CLOUD')
@@ -43,7 +48,6 @@ def test_cluster_ratio():
         im.set_clim(0,1)
         fig.colorbar(im)
         ax.set_title(name)
-
